delete example instrumented test move ReservationServiceRobolectricTest to appropriate location
